Compare covariates between intersections

Compare covariates between intersections

compare_between_intersections(
  data,
  intersect,
  test = kruskal.test,
  tests = list(),
  ignore = list(),
  ignore_mode_columns = TRUE,
  mode = "exclusive_intersection",
  ...
)

Arguments

data

a dataframe including binary columns representing membership in classes

intersect

which columns should be used to compose the intersection

test

the default test function; it is expected to accept formula and data parameters, and a list with p.value, statistic, and method

tests

a named list with tests for specific variables, overwriting the default test

ignore

a list with names of variables to exclude from testing

ignore_mode_columns

whether the membership columns and size columns for all modes should be ignored

mode

region selection mode; note that modes other than exclusive_intersection repeat observations in different test group, introducing dependencies. See get_size_mode() for accepted values.

...

passed to upset_data()