Barplot annotation of intersections sizes

Barplot annotation of intersections sizes

intersection_size(
  mapping = aes(),
  counts = TRUE,
  bar_number_threshold = 0.85,
  text_colors = c(on_background = "black", on_bar = "white"),
  text = list(),
  text_mapping = aes(),
  mode = "distinct",
  position = position_stack(),
  width = 0.9,
  ...
)

Arguments

mapping

additional aesthetics for geom_bar()

counts

whether to display count number labels above the bars

bar_number_threshold

if less than one, labels for bars height greater than this threshold will be placed on (not above) the bars

text_colors

a name vector of characters specifying the color when on_background and on_bar (see bar_number_threshold)

text

additional parameters passed to geom_text()

text_mapping

additional aesthetics for geom_text()

mode

region selection mode, defines which intersection regions will be accounted for when computing the size. See get_size_mode() for accepted values.

position

position passed to geom_bar()

width

bar width, by default set to 90%

...

Arguments passed on to ggplot2::geom_bar

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

just

Adjustment for column placement. Set to 0.5 by default, meaning that columns will be centered about axis breaks. Set to 0 or 1 to place columns to the left/right of axis breaks. Note that this argument may have unintended behaviour when used with alternative positions, e.g. position_dodge().

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

orientation

The orientation of the layer. The default (NA) automatically determines the orientation from the aesthetic mapping. In the rare event that this fails it can be given explicitly by setting orientation to either "x" or "y". See the Orientation section for more detail.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().