Color scale for Venn diagram

Color scale for Venn diagram

scale_color_venn_mix(
  data,
  sets = NULL,
  colors = c("red", "blue", "green"),
  na.value = "grey40",
  highlight = NULL,
  active_color = "orange",
  inactive_color = "NA",
  scale = scale_color_manual,
  ...
)

Arguments

data
a dataframe including binary columns representing membership in sets

sets
vector with names of columns representing membership in sets

colors
named list of colors for sets (one set=one color)

na.value
value for elements not belonging to any of the sets

highlight
which regions of the diagram to highlight

active_color
color for highlight

inactive_color
color for lack of highlight

scale
the base scale (default=scale_color_manual())

...
Arguments passed on to ggplot2::scale_color_manual

values

a set of aesthetic values to map data values to. The values will be matched in order (usually alphabetical) with the limits of the scale, or with breaks if provided. If this is a named vector, then the values will be matched based on the names instead. Data values that don't match will be given na.value.

aesthetics

Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. This can be useful, for example, to apply colour settings to the colour and fill aesthetics at the same time, via aesthetics = c("colour", "fill").

breaks

One of:

NULL for no breaks

waiver() for the default breaks (the scale limits)

A character vector of breaks

A function that takes the limits as input and returns breaks as output