Generate percentage label of the interestion/union sizes ratio

For use together with intersection_size or intersection_ratio

upset_text_percentage(digits = 0, sep = "")

Arguments

digits
How many digits to show when rounding the percentage?

sep
set to space (' ') if you prefer a whitespace between the number and the \% sign.

Examples

intersection_size(text_aes=ggplot2::aes_(label=upset_text_percentage()))

#> $aes
#> Aesthetic mapping: 
#> * `x` -> `intersection`
#> 
#> $geom
#> $geom[[1]]
#> geom_bar: width = NULL, na.rm = FALSE, orientation = NA
#> stat_count: width = NULL, na.rm = FALSE, orientation = NA
#> position_stack 
#> 
#> $geom[[2]]
#> mapping: y = ~ifelse(..count.. <= bar_number_threshold * max(..count..), ..count.., bar_number_threshold * ..count..), label = ~paste(round(intersection_size/union_size * 100, 0), "%", sep = ""), colour = ~ifelse(..count.. <= bar_number_threshold * max(..count..), "on_background", "on_bar") 
#> geom_text: parse = FALSE, check_overlap = FALSE, na.rm = FALSE
#> stat_count: na.rm = FALSE
#> position_identity 
#> 
#> $geom[[3]]
#> <ggproto object: Class ScaleDiscrete, Scale, gg>
#>     aesthetics: colour
#>     axis_order: function
#>     break_info: function
#>     break_positions: function
#>     breaks: waiver
#>     call: call
#>     clone: function
#>     dimension: function
#>     drop: TRUE
#>     expand: waiver
#>     get_breaks: function
#>     get_breaks_minor: function
#>     get_labels: function
#>     get_limits: function
#>     guide: FALSE
#>     is_discrete: function
#>     is_empty: function
#>     labels: waiver
#>     limits: NULL
#>     make_sec_title: function
#>     make_title: function
#>     map: function
#>     map_df: function
#>     n.breaks.cache: NULL
#>     na.translate: TRUE
#>     na.value: NA
#>     name: waiver
#>     palette: function
#>     palette.cache: NULL
#>     position: left
#>     range: <ggproto object: Class RangeDiscrete, Range, gg>
#>         range: NULL
#>         reset: function
#>         train: function
#>         super:  <ggproto object: Class RangeDiscrete, Range, gg>
#>     rescale: function
#>     reset: function
#>     scale_name: manual
#>     train: function
#>     train_df: function
#>     transform: function
#>     transform_df: function
#>     super:  <ggproto object: Class ScaleDiscrete, Scale, gg>
#> 
#> 
#> $highlight_geom
#> $highlight_geom[[1]]
#> geom_bar: width = NULL, na.rm = FALSE, orientation = NA
#> stat_count: width = NULL, na.rm = FALSE, orientation = NA
#> position_stack 
#> 
#>