Arrange points for Venn diagram

Arrange points for Venn diagram

arrange_venn(
  data,
  sets = NULL,
  radius = 1.5,
  max_iterations = 10,
  verbose = FALSE,
  outwards_adjust = 1.3,
  extract_sets = FALSE,
  extract_regions = FALSE,
  repeat_in_intersections = FALSE,
  starting_grid_size = "auto"
)

Arguments

data
a dataframe including binary columns representing membership in sets

sets
vector with names of columns representing membership in sets

radius
the radius of the circle

max_iterations
the maximal number of iterations

verbose
should debugging notes be printed?

outwards_adjust
the multiplier defining the distance from the centre

extract_sets
should only sets be extracted?

extract_regions
should all unique regions be extracted?

repeat_in_intersections
repeat intersection k times where k is the number of sets it belongs to?

starting_grid_size
the starting size of the grid for placement of elements