Prepare data for UpSet plots

Prepare data for UpSet plots

upset_data(
  data,
  intersect,
  min_size = 0,
  max_size = Inf,
  min_degree = 0,
  max_degree = Inf,
  n_intersections = NULL,
  keep_empty_groups = FALSE,
  warn_when_dropping_groups = TRUE,
  sort_sets = "descending",
  sort_intersections = "descending",
  sort_intersections_by = "cardinality",
  group_by = "degree",
  min_max_early = TRUE,
  union_count_column = "union_size",
  intersection_count_column = "intersection_size"
)

Arguments

data
a dataframe including binary columns representing membership in classes

intersect
which columns should be used to compose the intersection

min_size
minimal number of observations in an intersection for it to be included

max_size
maximal number of observations in an intersection for it to be included

min_degree
minimal degree of an intersection for it to be included

max_degree
maximal degree of an intersection for it to be included

n_intersections
the exact number of the intersections to be displayed; n largest intersections that meet the size and degree criteria will be shown

keep_empty_groups
whether empty sets should be kept (including sets which are only empty after filtering by size)

warn_when_dropping_groups
whether a warning should be issued when empty sets are being removed

sort_sets
whether to sort the rows in the intersection matrix (descending sort by default); one of: 'ascending', 'descending', FALSE

sort_intersections
whether to sort the columns in the intersection matrix (descending sort by default); one of: 'ascending', 'descending', FALSE

sort_intersections_by
the mode of sorting, the size of the intersection (cardinality) by default; one of: 'cardinality', 'degree', 'ratio'

group_by
the mode of grouping intersections; one of: 'degree', 'sets'

min_max_early
whether the min and max limits should be applied early (for faster plotting), or late (for accurate depiction of ratios)

union_count_column
name of the column to store the union size (adjust if conflicts with your data)

intersection_count_column
name of the column to store the intersection size (adjust if conflicts with your data)